requirements.txt
asgiref==3.7.1
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.3.0
drf-yasg==1.20.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
loginform==1.2.0
lxml==4.9.2
MarkupSafe==2.1.2
packaging==23.1
pymongo==3.12.0
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
sqlparse==0.2.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
